requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.13
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-docx==1.1.0
groq==0.4.1
numpy==1.24.3
torch==2.1.0
transformers==4.36.0
werkzeug==3.0.1